Added auto-resize for padding before footer
Use flexbox to create semi-sticky footer.

The footer should be fixed at the bottom of the page for pages whos
content does not reach the bottom. For pages whos content does reach the
bottom, the footer should be placed at the end of the content.

One thing I noticed was that there is no space between the footer and
the prev/next buttons when there is enough content to move the footer.

Made Cyber ed page headers/footers include the the base level div and body tags